refactor(convert): move replaced.rs tests into home module
